requirements.txt
fastapi
uvicorn
pandas
langchain
langchain-community
langchain-ollama
langchain-openai
openai
python-dotenv
